# openapi.yaml
...
openapi: 3.1.0
info:
  title: InferaDB Control
  description: |
    Control Plane for InferaDB providing self-service user authentication, organization management, and vault access control.

    ## Authentication

    Control uses two primary authentication methods:

    - **Session Cookie**: For web dashboard and interactive sessions. Set via `infera_session` cookie after login.
    - **JWT Bearer Token**: For programmatic access via client assertions (OAuth 2.0 JWT Bearer flow).

    ## Multi-Tenancy

    All resources are scoped to organizations. Users can belong to multiple organizations with different roles:

    - **Owner**: Full control, can transfer ownership
    - **Admin**: Manage members, vaults, clients, and teams
    - **Member**: Basic access, can view organization resources

    ## Rate Limiting

    API requests are rate-limited by IP address and user:

    - Login: 100 attempts/hour per IP
    - Registration: 5 per day per IP
    - Email verification: 5 tokens/hour
    - Password reset: 3 tokens/hour

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  version: 1.0.0
  contact:
    name: InferaDB Support
    url: https://inferadb.com/docs
    email: support@inferadb.com
  license:
    name: Business Source License 1.1
    url: https://github.com/inferadb/inferadb/blob/main/management/LICENSE.md
servers:
...
paths:
...
  /v1/users/emails:
    get:
      tags: [Emails]
      summary: List user emails
      operationId: listEmails
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Email list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserEmail"

    post:
      tags: [Emails]
      summary: Add email address
      description: Add additional email to user account (requires verification)
      operationId: addEmail
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                primary:
                  type: boolean
                  default: false
      responses:
        "201":
          description: Email added (verification sent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEmail"
        "409":
          description: Email already in use
  /v1/users/emails/{id}:
    patch:
      tags: [Emails]
      summary: Update email
      description: Set email as primary or update other properties
      operationId: updateEmail
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: snowflake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                primary:
                  type: boolean
      responses:
        "200":
          description: Email updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEmail"

    delete:
      tags: [Emails]
      summary: Delete email
      description: Remove email from account (cannot delete last email or unverified primary)
      operationId: deleteEmail
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: snowflake
      responses:
        "200":
          description: Email deleted
        "400":
          description: Cannot delete last email or primary email

  # Organization Endpoints
...
  /startupz:
    get:
      tags: [Health]
      summary: Startup probe
      description: |
        Kubernetes startup probe. Returns 200 once initialization is complete.
        Kubernetes will not send traffic until this succeeds.
      operationId: startupz
      responses:
        "200":
          description: Startup complete
        "503":
          description: Still initializing
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: infera_session
      description: Session cookie obtained from /v1/auth/login or /v1/auth/register
...
  schemas:
...
    UserEmail:
      type: object
      required: [id, user_id, email, is_primary, is_verified, created_at]
      properties:
        id:
          type: string
          format: snowflake
        user_id:
          type: string
          format: snowflake
        email:
          type: string
          format: email
        is_primary:
          type: boolean
        is_verified:
          type: boolean
        verified_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...